Move `Arena` type and `GetPairMut` trait into `arena` module
